+++ tests/testthat/test-partial.R
@@ -2,7 +2,7 @@ context("partial")
 
 test_that("dots are correctly placed in the signature", {
   out <- partialised_body(partial(runif, n = rpois(1, 5)))
-  exp <- quo((!!runif)(n = rpois(1, 5), ...))
+  exp <- quo(.fn(n = rpois(1, 5), ...))
   expect_identical(out, exp)
 })
 
@@ -136,12 +136,12 @@ test_that("`.lazy` still works", {
 test_that("`.first` still works", {
   scoped_options(lifecycle_disable_warnings = TRUE)
   out <- partialised_body(partial(runif, n = rpois(1, 5), .first = FALSE))
-  exp <- quo((!!runif)(..., n = rpois(1, 5)))
+  exp <- quo(.fn(..., n = rpois(1, 5)))
   expect_identical(out, exp)
 
   # partial() also works without partialised arguments
-  expect_identical(partialised_body(partial(runif, .first = TRUE)), quo((!!runif)(...)))
-  expect_identical(partialised_body(partial(runif, .first = FALSE)), quo((!!runif)(...)))
+  expect_identical(partialised_body(partial(runif, .first = TRUE)), quo(.fn(...)))
+  expect_identical(partialised_body(partial(runif, .first = FALSE)), quo(.fn(...)))
 })
 
 test_that("`...f` still works", {